Clarify config loading pt. 5
* Renamed user_lang_config to merged_lang_config, as it
includes the default language config.

* Moved the functions of helix_loader::config (relate only
to langage config) to helix-loader/lib.rs, and removed
the file. Mostly to aid in comparison in how the config
is loaded.

* Moved the functions of helix_core::config to
helix_core::syntax. This way they're loaded the same way as
the config.toml. Forces a consistency initialsing language
config with LanguageConfigurations::default() and ::merged()
respectively. helix_core::config.rs was then removed.
