+++ helix-core/src/lib.rs
@@ -3,7 +3,6 @@ pub use encoding_rs as encoding;
 pub mod auto_pairs;
 pub mod chars;
 pub mod comment;
-pub mod config;
 pub mod diagnostic;
 pub mod diff;
 pub mod graphemes;
