+++ helix-loader/src/grammar.rs
@@ -191,7 +191,7 @@ pub fn build_grammars(target: Option<String>) -> Result<()> {
 // merged. The `grammar_selection` key of the config is then used to filter
 // down all grammars into a subset of the user's choosing.
 fn get_grammar_configs() -> Result<Vec<GrammarConfiguration>> {
-    let config: Configuration = crate::config::merged_lang_config()
+    let config: Configuration = crate::merged_lang_config()
         .context("Could not parse languages.toml")?
         .try_into()?;
