+++ helix-term/src/main.rs
@@ -140,13 +140,15 @@ FLAGS:
         Err(err) => return Err(Error::new(err)),
     };
 
-    let syn_loader_conf = helix_core::config::user_syntax_loader().unwrap_or_else(|err| {
+    let syn_loader_conf =  helix_loader::default_lang_config().try_into().unwrap_or_else(|err| {
         eprintln!("Bad language config: {}", err);
         eprintln!("Press <ENTER> to continue with default language config");
         use std::io::Read;
         // This waits for an enter press.
         let _ = std::io::stdin().read(&mut []);
-        helix_core::config::default_syntax_loader()
+        helix_loader::default_lang_config()
+            .try_into()
+            .expect("Could not serialize built-in languages.toml")
     });
 
     // TODO: use the thread local executor to spawn the application task separately from the work pool
